requirements.txt
matplotlib
numpy
scipy
pytorch==0.4.1
torchvision==0.2.0
tqdm==4.11.2
protobuf